requirements.txt
Pillow==9.2.0
pytorch_lightning==1.3.5
torchmetrics==0.3.2
torchvision==0.11.3
tqdm==4.64.0
